Face Detection/pyproject.toml
[project]
name = "Face Detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cmake>=4.0.3",
    "customtkinter>=5.2.2",
    "dlib>=20.0.0",
    "face-recognition>=1.3.0",
    "imutils>=0.5.4",
    "matplotlib>=3.10.3",
    "mediapipe>=0.10.14",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pygame>=2.6.1",
    "python-vlc>=3.0.21203",
    "seaborn>=0.13.2",
    "torch>=2.7.1",
    "ultralytics>=8.3.168",
]
